backend/requirements.txt
# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0

# AI/ML Libraries
langchain==0.0.340
langchain-community==0.0.1
langchain-huggingface==0.0.1
transformers==4.35.0
accelerate==0.24.1
bitsandbytes==0.41.1
torch==2.1.0
sentence-transformers==2.2.2
chromadb==0.4.18

# Document Processing
pypdf==3.17.0
python-docx==1.1.0
pillow==10.1.0
python-magic-bin==0.4.14

# Data Processing
pandas==2.1.3
numpy==1.24.3
scikit-learn==1.3.2

# Visualization
matplotlib==3.8.2
plotly==5.18.0
kaleido==0.2.1

# Database
supabase==1.1.1
sqlalchemy==2.0.23
alembic==1.12.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.1.2

# Hugging Face
huggingface-hub==0.19.4
huggingface-hub[cli]==0.19.4

# File Operations
aiofiles==23.2.1
python-magic==0.4.27

# Utilities
python-dateutil==2.8.2
pytz==2023.3
typing-extensions==4.8.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring & Logging
psutil==5.9.6
python-json-logger==2.0.7
structlog==23.2.0

# Optional (for production)
# redis==5.0.1
# celery==5.3.4
# requests==2.31.0
# httpx==0.25.1
# email-validator==2.1.0